Scripteable objects
Player settings toegevoegd
player animations zijn nog een beetje buggy, voornamelijk omdat "a & d"  hetzelfde doen, dus daar een check aan toevoegen.
