+++ PrjContextII/Assets/_PrjContextII/Scenes/SampleScene.unity
@@ -1778,6 +1778,7 @@ MonoBehaviour:
   computerLayermask:
     serializedVersion: 2
     m_Bits: 8
+  PS: {fileID: 11400000, guid: f4f34279cd32fd247b889e95bf75b124, type: 2}
 --- !u!136 &1244257128
 CapsuleCollider:
   m_ObjectHideFlags: 0
