+++ PrjContextII/Assets/_PrjContextII/ScripteableObjects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1a565744cdbd734489574167c6926b56
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
