+++ PrjContextII/Assets/_PrjContextII/ScripteableObjects/PlayerSettings.asset
@@ -1,0 +1,21 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b1cdc61372e5fbe46b4923bac282598c, type: 3}
+  m_Name: PlayerSettings
+  m_EditorClassIdentifier: 
+  speed: 2
+  computerLayerMask:
+    serializedVersion: 2
+    m_Bits: 0
+  MainCam: {fileID: 0}
+  rb: {fileID: 0}
+  anim: {fileID: 0}
