+++ PrjContextII/Assets/_PrjContextII/ScripteableObjects/PlayerSettings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f4f34279cd32fd247b889e95bf75b124
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
