+++ PrjContextII/Assets/_PrjContextII/Scripts/Player/ComputerInteract.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6e1c0d9bf1b46fe4daff808ec2b41f32
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
