+++ PrjContextII/Assets/_PrjContextII/Scripts/Player/PlayerMovementFree.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 44efde877d7fc46488ddf7cee4c2116d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
