+++ PrjContextII/Assets/_PrjContextII/Scripts/ScripteableObjects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 03a7555a529348549adff121fd697cb7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
