+++ PrjContextII/Assets/_PrjContextII/Scripts/ScripteableObjects/PlayerSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b1cdc61372e5fbe46b4923bac282598c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
